Remove arbitrary stop() calls when a YAML file is stored in a subdirectory
This is an absolutely disgusting hack. These stop()s were put in because in an ideal world you could have the yaml file in a subdirectory and its paths could be relative to that subdir. This doesn't resolve that - all paths still must be relative to the top-level project dir. But it at least allows files to be organized within a subdirectory.